object and Vect2

+++ include/P_types.hpp
@@ -2,5 +2,5 @@
 
 #include <cstddef>
 
-using pos_type = int;
-using time_type = size_t;
+using pos_type = double;
+using time_type = double;
